requirements.txt
streamlit>=1.35.0,<2.0
pandas>=2.0.0,<3.0
google-analytics-data>=0.18.0,<1.0
google-analytics-admin>=0.22.0,<1.0
google-auth>=2.29.0,<3.0
google-auth-oauthlib>=1.2.0,<2.0
google-api-python-client>=2.130.0,<3.0
google-generativeai>=0.7.0,<1.0
python-slugify>=8.0.0,<9.0
pypdf>=4.0.0,<5.0
openpyxl>=3.1.0,<4.0
xlrd>=2.0.0,<3.0
